\name{golden}
\alias{golden}
\title{Metode Golden Ratio}
\description{Mencari nilai minimizer suatu fungsi dengan metode golden ratio}
\usage{golden(f,a,b)}
\arguments{
  \item{f}{fungsi}
  \item{a}{batas interval atas}
  \item{b}{batas interval bawah}
}
\details{Fungsi ini digunakan untuk mencari nilai minimizer suatu fungsi dengan metode golden ratio}
\value{Nilai Minimizer}
\references{Utari,2023,Modul Komputasi Statistika dan Aplikasinya dengan R,UII,Yogyakarta}
\author{Wulan,Najla,Dhika,Alya,Renata}
\examples{\dontrun{golden(f,0,4)}}
