\name{ujisampel}
\alias{ujisampel}
\title{Uji Hipotesis Satu Sampel t}
\description{Menguji pernyataan yang diajukan tentang populasi berdasarkan data dari satu sampel}
\usage{ujisampel(data,mu)}
\arguments{
  \item{data}{data}
  \item{mu}{mean populasi}
}
\details{Fungsi ini digunakan untuk Menguji pernyataan yang diajukan tentang populasi berdasarkan data dari satu sampel}
\value{Keputusan Uji Hipotesis}
\references{}
\author{Wulan,Najla,Dhika,Alya,Renata}
\examples{\dontrun{ujisampel(data,mu)}}
